Removed LiveData matchers
